refactor : MemoListModel 역할 분배
