+++ CloudNotes/CloudNotes/View/MemoListViewController.swift
@@ -56,10 +56,9 @@ class MemoListViewController: UIViewController {
     }
     
     @objc private func didTapAdd() {
-        memoListViewModel.createMemoData()
-        
+        memoListViewModel.coreDataMemoModel.createMemoData()
+        memoListViewModel.getAllMemoData()
         if UITraitCollection.current.horizontalSizeClass == .compact {
-            self.memoListViewModel.setupEditingMemo(index: 0)
             memoDetailViewDelegate?.setupDetailText(data: memoListViewModel.readMemo(index: 0))
             guard let detail = memoDetailViewDelegate as? MemoDetailViewController else { return }
             self.navigationController?.pushViewController(detail, animated: true)
@@ -70,7 +69,8 @@ class MemoListViewController: UIViewController {
     
     @objc private func deleteMemo() {
         guard let memoData = self.memoListViewModel.editingMemo else { return }
-        self.memoListViewModel.deleteMemoData(data: memoData)
+        self.memoListViewModel.coreDataMemoModel.deleteMemoData(data: memoData)
+        self.memoListViewModel.getAllMemoData()
         reloadTableView()
     }
     
@@ -79,11 +79,11 @@ class MemoListViewController: UIViewController {
         guard let memoData = self.memoListViewModel.editingMemo else { return }
         let subString = textData.subString(target: textData, point: "\n")
         if subString.count == 2 {
-            self.memoListViewModel.updateMemoData(titleText: subString[0],
+            self.memoListViewModel.coreDataMemoModel.updateMemoData(titleText: subString[0],
                                                   bodyText: subString[1],
                                                   data: memoData)
         } else {
-            self.memoListViewModel.updateMemoData(titleText: textData, bodyText: "", data: memoData)
+            self.memoListViewModel.coreDataMemoModel.updateMemoData(titleText: textData, bodyText: "", data: memoData)
         }
         reloadTableView()
     }
@@ -117,7 +117,7 @@ extension MemoListViewController: UITableViewDelegate, UITableViewDataSource {
         }
         let memoData = memoListViewModel.readMemo(index: indexPath.row)
         cell.configureCell(memoData: memoData,
-                           stringLastModified: memoListViewModel.convertDate(date: memoData.lastModified))
+                           stringLastModified: memoListViewModel.coreDataMemoModel.converter.convertDate(date: memoData.lastModified))
         
         return cell
     }
